tests/wpt/web-platform-tests/css/css-properties-values-api/animate-invalid.html

<!DOCTYPE html>
<title>Do not crash when animating to unresolved var()</title>
<link rel="help" href="https://drafts.css-houdini.org/css-properties-values-api/#dom-css-registerproperty">
<link rel="help" href="https://crbug.com/1185524">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
  promise_test(async function(){
    CSS.registerProperty({
      name: '--x',
      syntax: '<number>',
      initialValue: '1',
      inherits: false
    });
    let animation = target.animate({'--x': [ 'var(--unknown)']}, 100);
    await animation.ready;
    assert_equals(getComputedStyle(target).getPropertyValue('--x'), '1');
  }, 'Do not crash when animating to unresolved var()');
</script>
